ADD: tests for the forcefields
